python/rest-simple/requirements/base.txt
pytz==2022.2.1  # https://github.com/stub42/pytz
whitenoise==6.2.0  # https://github.com/evansd/whitenoise
requests==2.28.1

# Django
# ------------------------------------------------------------------------------
Django==4.1
djangorestframework~=3.13.1
django-cors-headers==3.13.0  # https://github.com/adamchainz/django-cors-headers
# psycopg2==2.9.3  # https://github.com/psycopg/psycopg2
gunicorn==20.1.0
drf-yasg==1.21.3
